+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 - Theme template and theme handler.
 - Support customized theme on mixin.
 - Theme handler adapt system brightness change.
+- Remove unnecessary encapsulations and tidy comments.
+- Inherit on change trigger (experimental).
 
 ## 0.3.0